AgenticAI/system_rca_client_server_mcp/requirements.txt
# MCP protocol
mcp>=1.2.0

# async runtime required by MCP
anyio>=4.3.0
typing_extensions>=4.10.0

# LLM client
openai>=1.14.0

# server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# utilities
python-dotenv>=1.0.1
httpx>=0.27.0

# optional but recommended
pydantic>=2.6.0
typing_inspection>=0.4.0
